refactor: folder structure and modularised
