+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0-semantically-released",
   "description": "cli for xmanscript tools",
   "type": "module",
+  "module": "es2020",
   "scripts": {
     "dev": "tsup --watch",
     "build": "tsup",
@@ -34,6 +35,7 @@
   },
   "homepage": "https://github.com/laxmanpokhrel/xmanscript-cli#readme",
   "dependencies": {
+    "@types/inquirer": "^9.0.7",
     "chalk": "^5.3.0",
     "child-process": "^1.0.2",
     "cross-env": "^7.0.3",
