+++ src/tasks/backend.ts
@@ -1,0 +1,3 @@
+export default async function backend() {
+  console.log("Yet to create boilerplate for backend.");
+}
