+++ src/types.d.ts
@@ -1,3 +1,2 @@
-// declare module "child_process";
 declare module "fs";
 declare module "util";
