+++ src/utils/index.ts
@@ -1,0 +1,6 @@
+import extractFlags from "./extractFlags";
+import getRepoName from "./getRepoName";
+import getGitAuthor from "./getGitAuthor";
+import configurePackageJson from "./configurepackagejson";
+
+export { extractFlags, getRepoName, getGitAuthor, configurePackageJson };
